backend/local-api/requirements.txt
flask
flask-cors
requests
python-dotenv
numpy
scikit-learn
joblib
paho-mqtt
